fix::ambulance: fix inner private class that prevent jackson binding to work properly
