+++ code/api/api/src/main/java/com/decathlon/ara/service/dto/request/ExecutedScenarioHistoryInputDTO.java
@@ -60,46 +60,4 @@ public class ExecutedScenarioHistoryInputDTO {
     public String getRunTypeCode() {
         return runTypeCode;
     }
-
-    private class ExecutedScenarioHistoryDuration {
-        private int value;
-        private ExecutedScenarioHistoryDurationType type;
-
-        private Optional<Period> getDuration() {
-            if (type == null || value < 1) {
-                return Optional.empty();
-            }
-            Period period;
-            switch (type) {
-                case DAY:
-                    period = Period.ofDays(value);
-                    break;
-                case WEEK:
-                    period = Period.ofWeeks(value);
-                    break;
-                case MONTH:
-                    period = Period.ofMonths(value);
-                    break;
-                case YEAR:
-                    period = Period.ofYears(value);
-                    break;
-                default:
-                    period = Period.ZERO;
-            }
-            return Optional.of(period);
-        }
-
-        public int getValue() {
-            return value;
-        }
-
-        public ExecutedScenarioHistoryDurationType getType() {
-            return type;
-        }
-    }
-
-    private enum ExecutedScenarioHistoryDurationType {
-        DAY, WEEK, MONTH, YEAR
-    }
-
 }
